requirements.txt
beautifulsoup4==4.10.0
python-dotenv==0.19.2
requests==2.27.1
spotipy==2.19.0
